requirements.txt
python-telegram-bot==20.6
pymongo[srv]
python-dotenv
